+++ modules/stake/handler.go
@@ -78,6 +78,7 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) (validatorUpdates []abci.Valid
 			tags.DstValidator, []byte(dvvTriplet.ValidatorDstAddr.String()),
 		))
 	}
+	k.UpdateMetrics(ctx)
 	return
 }