+++ server/config/config.go
@@ -14,6 +14,9 @@ const (
 type BaseConfig struct {
 	// Tx minimum fee
 	MinFees        string `mapstructure:"minimum_fees"`
+
+	// Enable invariant check, ignore this flag on testnet
+	CheckInvariant bool `mapstructure:"check_invariant"`
 }
 
 // Config defines the server's top level configuration
@@ -35,5 +38,5 @@ func (c *Config) MinimumFees() sdk.Coins {
 
 // DefaultConfig returns server's default configuration.
 func DefaultConfig() *Config {
-	return &Config{BaseConfig{MinFees: defaultMinimumFees}}
+	return &Config{BaseConfig{MinFees: defaultMinimumFees, CheckInvariant: false}}
 }