+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.12.0
+
+*February 1st, 2019*
+
+[iris] Refactor and add more metrics for monitor
+[iris] Improve the config of Invariant Checking
+
+[iriscli] Add share percent in lcd unbond and redelegate 
+[iriscli] Fix tx search bug by page and size query parameter name
+
+[tendermint] Fix DynamicVerifier for large validator set changes
+[tendermint] Optimize txs search
+
+
 ## 0.11.0
 
 *January 25th, 2019*